+++ sdk/search/azure-search-documents/azure/search/documents/indexes/_internal/__init__.py
@@ -10,4 +10,4 @@ from ._index import (  # pylint: disable=unused-import
 from ._search_index_client import SearchIndexClient  # pylint: disable=unused-import
 from ._search_indexer_client import SearchIndexerClient  # pylint: disable=unused-import
 
-from . import edm  # pylint: disable=unused-import
+from . import SearchFieldDataType  # pylint: disable=unused-import