+++ sdk/search/azure-search-documents/azure/search/documents/indexes/models/__init__.py
@@ -28,7 +28,7 @@ from .._internal import (
     ComplexField,
     SearchableField,
     SimpleField,
-    edm,
+    SearchFieldDataType,
 )
 from .._internal._generated.models import (
     AnalyzeRequest,
@@ -209,5 +209,5 @@ __all__ = (
     "UniqueTokenFilter",
     "WebApiSkill",
     "WordDelimiterTokenFilter",
-    "edm",
+    "SearchFieldDataType",
 )