+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_receipt_from_url.py
@@ -263,8 +263,7 @@ class TestReceiptFromUrl(FormRecognizerTest):
         self.assertEqual(receipt.merchant_name.value, 'Bilbo Baggins')
         self.assertEqual(receipt.merchant_phone_number.value, '+15555555555')
         self.assertEqual(receipt.subtotal.value, 300.0)
-        # TODO: revert after service side fix
-        self.assertIsNotNone(receipt.total.value)
+        self.assertEqual(receipt.total.value, 100.0)
         self.assertEqual(receipt.page_range.first_page, 1)
         self.assertEqual(receipt.page_range.last_page, 1)
         self.assertFormPagesHasValues(receipt.pages)