+++ sdk/search/azure-search-documents/azure/search/documents/indexes/_internal/aio/__init__.py
@@ -1,0 +1,11 @@
+# ------------------------------------
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+# ------------------------------------
+from ._search_indexer_client import SearchIndexerClient
+from ._search_index_client import SearchIndexClient
+
+__all__ = (
+    "SearchIndexerClient",
+    "SearchIndexClient",
+)