+++ sdk/search/azure-search-documents/tests/test_index_field_helpers.py
@@ -3,24 +3,24 @@
 # Licensed under the MIT License.
 # ------------------------------------
 
-from azure.search.documents import ComplexField, SearchableField, SimpleField, edm
+from azure.search.documents.indexes.models import ComplexField, SearchableField, SimpleField, SearchFieldDataType
 
 def test_edm_contents():
-    assert edm.String == "Edm.String"
-    assert edm.Int32 == "Edm.Int32"
-    assert edm.Int64 == "Edm.Int64"
-    assert edm.Double == "Edm.Double"
-    assert edm.Boolean == "Edm.Boolean"
-    assert edm.DateTimeOffset == "Edm.DateTimeOffset"
-    assert edm.GeographyPoint == "Edm.GeographyPoint"
-    assert edm.ComplexType == "Edm.ComplexType"
-    assert edm.Collection("foo") == "Collection(foo)"
+    assert SearchFieldDataType.String == "Edm.String"
+    assert SearchFieldDataType.Int32 == "Edm.Int32"
+    assert SearchFieldDataType.Int64 == "Edm.Int64"
+    assert SearchFieldDataType.Double == "Edm.Double"
+    assert SearchFieldDataType.Boolean == "Edm.Boolean"
+    assert SearchFieldDataType.DateTimeOffset == "Edm.DateTimeOffset"
+    assert SearchFieldDataType.GeographyPoint == "Edm.GeographyPoint"
+    assert SearchFieldDataType.ComplexType == "Edm.ComplexType"
+    assert SearchFieldDataType.Collection("foo") == "Collection(foo)"
 
 class TestComplexField(object):
     def test_single(self):
         fld = ComplexField(name="foo", fields=[])
         assert fld.name == "foo"
-        assert fld.type == edm.ComplexType
+        assert fld.type == SearchFieldDataType.ComplexType
 
         assert fld.sortable is None
         assert fld.facetable is None
@@ -34,7 +34,7 @@ class TestComplexField(object):
     def test_collection(self):
         fld = ComplexField(name="foo", fields=[], collection=True)
         assert fld.name == "foo"
-        assert fld.type == edm.Collection(edm.ComplexType)
+        assert fld.type == SearchFieldDataType.Collection(SearchFieldDataType.ComplexType)
 
         assert fld.sortable is None
         assert fld.facetable is None
@@ -47,9 +47,9 @@ class TestComplexField(object):
 
 class TestSimplexField(object):
     def test_defaults(self):
-        fld = SimpleField(name="foo", type=edm.Double)
+        fld = SimpleField(name="foo", type=SearchFieldDataType.Double)
         assert fld.name == "foo"
-        assert fld.type == edm.Double
+        assert fld.type == SearchFieldDataType.Double
         assert fld.retrievable == True
         assert fld.sortable == False
         assert fld.facetable == False
@@ -63,9 +63,9 @@ class TestSimplexField(object):
 
 class TestSearchableField(object):
     def test_defaults(self):
-        fld = SearchableField(name="foo", type=edm.Collection(edm.String))
+        fld = SearchableField(name="foo", type=SearchFieldDataType.Collection(SearchFieldDataType.String))
         assert fld.name == "foo"
-        assert fld.type == edm.Collection(edm.String)
+        assert fld.type == SearchFieldDataType.Collection(SearchFieldDataType.String)
         assert fld.retrievable == True
         assert fld.sortable == False
         assert fld.facetable == False