+++ sdk/search/azure-search-documents/CHANGELOG.md
@@ -5,7 +5,8 @@
 **Breaking Changes**
 
 - Reorganized `SearchServiceClient` into `SearchIndexClient` & `SearchIndexerClient`    #11507
-- Split searchindex.json and searchservice.json models and operations into separate namespaces #11508
+- Splitted searchindex.json and searchservice.json models and operations into separate namespaces #11508
+- Renamed `edm` to `SearchFieldDataType`    #11511
 
 ## 1.0.0b3 (2020-05-04)